changed scale Factor from int to double
git-svn-id: file:///srv/local/data/cleanRepo/trunk@8461 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/org/matsim/locationchoice/facilityload/EventsToFacilityLoad.java
@@ -43,7 +43,7 @@ public class EventsToFacilityLoad implements ActivityStartEventHandler, Activity
 	private TreeMap<Id, FacilityPenalty> facilityPenalties;
 	private final static Logger log = Logger.getLogger(EventsToFacilityLoad.class);
 
-	public EventsToFacilityLoad(final ActivityFacilities facilities, int scaleNumberOfPersons,
+	public EventsToFacilityLoad(final ActivityFacilities facilities, double scaleNumberOfPersons,
 			TreeMap<Id, FacilityPenalty> facilityPenalties) {
 		super();
